# admin-api/api/employee/types.go
package employee

type CreateEmployeeRequest struct {
	Name     string `json:"name"`
	Password string `json:"password"`
	Avatar   string `json:"avatar"`
}

type LoginRequest struct {
	Name     string `json:"name"`
	Password string `json:"password"`
}

type LoginResponse struct {
	Token string `json:"token"`
}

type IdUri struct {
	Id uint `uri:"id"`
}